requirements.txt
dj-database-url==0.4.2
Django==1.11.6
djangorestframework-jsonapi==2.2.0
djangorestframework_simplejwt==3.2
djangorestframework==3.6.4
psycopg2==2.7.3.2
pytz==2017.2
